refactor: extract shared background renderable
