+++ core/src/main/java/com/p1_7/game/entities/BackgroundImage.java
@@ -1,7 +1,5 @@
-package com.p1_7.game.scenes.settings;
+package com.p1_7.game.entities;
 
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.graphics.Texture;
 import com.p1_7.abstractengine.render.IDrawContext;
 import com.p1_7.abstractengine.render.IRenderable;
 import com.p1_7.abstractengine.transform.ITransform;
@@ -9,17 +7,23 @@ import com.p1_7.game.Settings;
 import com.p1_7.game.core.Transform2D;
 import com.p1_7.game.platform.GdxDrawContext;
 
-final class SettingsBackground implements IRenderable {
+/**
+ * Shared full-screen background renderable that delegates texture ownership
+ * to the engine asset pipeline.
+ */
+public final class BackgroundImage implements IRenderable {
 
     private final Transform2D transform;
     private final String assetPath;
-    private final Texture texture;
 
-    SettingsBackground(String assetPath) {
+    public BackgroundImage(String assetPath) {
         this.assetPath = assetPath;
-        this.texture = new Texture(Gdx.files.internal(assetPath));
-        this.texture.setFilter(Texture.TextureFilter.Linear, Texture.TextureFilter.Linear);
-        this.transform = new Transform2D(0f, 0f, Settings.getWindowWidth(), Settings.getWindowHeight());
+        this.transform = new Transform2D(
+            0f,
+            0f,
+            Settings.getWindowWidth(),
+            Settings.getWindowHeight()
+        );
     }
 
     @Override
@@ -38,11 +42,5 @@ final class SettingsBackground implements IRenderable {
         gdxCtx.drawTexture(assetPath,
             transform.getPosition(0), transform.getPosition(1),
             transform.getSize(0), transform.getSize(1));
-    }
-
-    void dispose() {
-        if (texture != null) {
-            texture.dispose();
-        }
     }
 }
